Change DoubleSpinner Decimals to byte type
Change the data type of the Decimals property to be a byte. This is a breaking change, technically, but I previously disallowed values above 15 anyway.

I also updated the default value to be 15. This way, decimal values of the full allowable range are supported out of the box, rather than forcing them to be rounded to 2 places. Trailing zeros are not included.

(Lastly, I finished my thoughts I had on the CheckBox control that I don't think I finished filling out in a looooooooong time.)