Remove obsolete method from MessageDialog
Took me soooo long to finally do this. I had marked it as obsolete so many versions ago, and I finally actually did it! This will also make it finally a bit easier to use this method without needing to specify the buttonDisplay value anymore.

This is a breaking change.