Create TimeSpanEditor, start on DateTimeEditor
Created a TimeSpanEditor control to use with the PropertyList; I think it's mostly complete (although I'll need to start testing it at some point).

I also created a DateTimeEditor, but this isn't 100% complete. Having to account for the potential of a DateOnly value kind of makes the DateTimeEditor a bit trickier to put together. So I might redo how I've designed part of it so far...

Anyway, these aren't fully added into the PropertyList itself yet, that'll be in a future commit. But we're on our way to including support for a common data type that I wasn't supporting before!