TimeSpinner - Fixes with Use12HourClock property
Now the UI is refreshed when this property is changed. Also, I've added in cases/support for when it is between noon and 1 PM or midnight and 1 AM, to display the correct value.

Right now, the TimeSpinner does not support times being entered in using the AM or PM designators, so that might require some further fixing later.