Add properly named Async method for WriteToFile
The existing WriteToFile method was actually an async method, but I didn't actually name it "WriteToFileAsync". Now it is properly named that, and I also added a new WriteToFile function that isn't async at all.

While in most cases, I'd expect everyone to just use the WriteToFileAsync (in which case they'll need to update their code to use the current version), I figured I'd add this in in case it's needed by someone somewhere.