Add MenuChanged, MenuOpening, MenuOpened to SplitButton
Similar to the MenuButton changes I just did!

One thing I didn't note is that for both the MenuButton and SplitButton, I actually didn't properly mark MenuClosed as an event before, and so now I did! I was originally also going to have them be as RoutedEventHandler, but I decided not to - to avoid the misperception that these would be routed events when they are not.