Finish setting up IntegerSpinner and LongSpinner
Templates have been made for both of these, and now both should be up and running, ready to go!

With the work that was already done to get these two spinners up and running, it should be easier now to get the remaining spinner classes (DoubleSpinner, TimeSpinner, StringSpinner) added in!