Add Spinners to controls file, testing window
Added in the spinners file to the main controls resource dictionary, so now they're able to be directly used!

Also added on a DoubleSpinner to the MainWindow, so that I can start to test out its functionality and handling with the Decimals property.

With the MainWindow, I also created a ColorScheme property that I could connect to, to showcase/test that ColorScheme binding was working. Eventually when the MainWindow is changed into a FlatWindow, that should have that already included in there.