Add TimeSpinner control
With my previous issues I had with my spinner controls out of the way, now I've added on the newly-built TimeSpinner control.

Minus the StringSpinner control (which won't be hard to set up, but will take some extra time and thought), all of the other spinner controls are now here!