TimeSpinner - add parsing for 12-hour times
Added parsing for 12-hour times (i.e. times ending with AM and PM), by utilizing the time parsing support from DateTime. This parsing will be a bit culture-dependent, so it may not consistently always parse the correct time, but this should work well enough for what's needed here, and I can always try to mess with this more in the future.

Also updated and fixed some of the comments at the top of the file.