Move CaptionType to own file, add extension methods
Moved CaptionType to its own file, so that it's a bit cleaner/easier to access the enum. Also added on some extension methods to quickly add or remove the Minimize and Maximize buttons. Yay for convenience!

Finally, added on a function to FlatWindow to automatically update the CaptionDisplayType if the ResizeMode value is changed.