Fix errors/warnings with build, update SourceLink
I'm using a different computer at the moment (actually, my old computer that a lot of this library was built on), and it seems unhappy about the way that I've done the compiler directives for WindowChromeWrapper. So I changed it to be #if NET40, rather than testing that it's *not* NET40, and it seems happy lol.

Also fixed or disabled a warning about stream.Read - I'm not going to fix it in code that I won't be supporting, and in the code I am supporting, the suggested stream.ReadExactly is only present in .NET 7 and later. Soooo yeah lol.

Also updated the GitHub SourceLink to 8.0.0. It says on the repo page that SourceLink is now included in the .NET 8 SDK, but the issue is that I'm targetting a lot of old versions before .NET 8, so I'll keep this NuGet package around.