Backport CursorEditor from 2.0
There's nothing in here that isn't relying upon my other SSUI 2.0 work, so I figure I may as well backport this so that it can be used here.

This allows the Cursor property in WPF controls to be updated and changed. It also supports reading cursors from a file. As part of this, I had to create an extension method that allowed me to pull in the corresponding Cursor type from the CursorType enum, so that was also backported to here.