Fix arrows not aligning to the right in ContextMenu
When you have a ContextMenu where its items have subitems, the arrows indicating the subitems weren't correctly aligned to the right. Turns out, the ContextMenu wasn't set correctly to actually line up the arrows. Not really a debate about this being an oversight, but the reality is that I frankly try to not delve much into the template code for Menu or ContextMenu lol.