New RectEditDialog, add ColorScheme to RectEdit
Created a new RectEditDialog, I guess I just straight up deleted it the previous time I tried and ran into a problem. Have no problems this time though! I think I'm a bit more loose with the size constraints of the RectEdit control in this new dialog, so maybe that helps? I'm not sure.

Anyway, I also made some updates to the RectEdit control to also add in the ColorScheme property, and make sure everything is looking good there.

Soon, I'll just have to go and create a new RectEditor for the PropertyList - or just find the old RectEditor in that previous commit and re-add it here lol.