Create RectEditor for PropertyList, get it set up
I've created the RectEditor for the PropertyList, to be able to edit Rect objects. That's not a redundant sentence in any way lol.

The basics and core stuff are all here, in theory, but there are a few things I'm realizing I need to do here.

The value "Rect.Empty" is not the number I was expecting it to be, and so I'm going to need to more properly handle that in the RectEdit control. Also, switching between the measure modes in the RectEdit control doesn't update the other spinners' values, so that'll need to be fixed too. Finally, a better output or representation in the RectEditor than just the ToString output of the Rect might also be helpful here - granted, that's not as much of a priority.