Add support for access keys in buttons, CheckBox
Backported access key support from 2.0... not really much of a backport, as it is just enabling a feature that I already enabled in the main (2.0) branch and now I'll just enable here.

Now you can use the underscore in your XAML to indicate an access key for pressing that button. Such a standard feature in WPF, but I didn't know how easy it was to add support for lol.

Well, better late than never, I guess!