Rename WriteToFile to WriteToFileAsync
Not quite just that. I've made the .NET 4 (non-async) version of WriteToFile available to all versions of the library now, and then for .NET 4.5 and later, I renamed the async version to WriteToFileAsync. So now almost everyone gets both!

This is like the only async method in the entire library, I think lol.

This is a breaking change.