Marking TabItem.IsVisible as obsolete
I will be removnig this from TabItem in version 2.0. I didn't realize when I added it in that the Control equivalent was primarily just used as a trigger source property. These kind of properties aren't really something I've built or had in mind for TabItem, so no point in having this redundant property here when Visibility is the preferable option.