Add brushes, ColorSchemeChanged to TabControl
I realized while putting together these changes for 1.9.9 that TabControl does not use brushes at all, but instead pulls all its brushes from the ColorScheme property. I want to change this in case we have people who use TabControl but don't want to set up a ColorScheme. I won't be able to make it a perfect system right now, but I'll at least aim to make some improvements here.

In this commit, I've added functionality to the existing BorderBrush, Background, and Foreground brushes. I've also added in the ContentAreaBackground brush which can be used to color the content area (it was otherwise just transparent this whole time lol). I also added in the ColorSchemeChanged event, which apparently I never put in for TabControl.