Add tab brush properties to TabControl
Added in the tab brush properties that I'll use later for the TabControl to control the brushes of the TabDisplayItem. As of right now, these properties don't do anything. I also added in the SelectedTabBackground property to the TabDisplayItem.

I've set up a number of the pieces I currently need, so in a commit coming soon, I'll make the adjustments of how brushes are transferred to the TabDisplayItem, and stop using ColorScheme.

Also in this commit, I removed the check for the PART_Content control in the template, since I actually never used it for anything.