Improvements to GradientBar, GradientStopItem
I didn't realize how much I didn't surface to be edited or modified, making this very limited how it can be used in other scenarios.

I've added a lot of methods to GradientBar to allow it to be controlled via these methods - doing things like changing the selection, adding/removing stops, and editing the currently selected stop.

I added and fixed brushes in the GradientStopItem, and exposed access to them via the GradientBar, so that their appearance can be changed.

Now, the GradientBar should be much more useful for others!