Remove a few items from IClickSelectableControl
Trying to strike a balance between adding in some properties and values for convenience, and redeclaring the entire FrameworkElement class in this interface.

With the XML comment, I made clear that it's expected that this interface will be applied to a FrameworkElement.

For VerifyAccess and CheckAccess, these are DispatcherObject methods that Microsoft decided to hide from IntelliSense... so I guess I'll hide these as well. By just not including these.