Implement BorderSelectedBrush in CheckBox
After having removed it in a previous commit, now I've reversed course and re-added and fully implemented the previously non-functional brush.

I have the brush's properties set to be the same as the standard BorderBrush, to continue to mimic the existing visual appearance from previous versions where there wasn't a selected brush. But now there is!