Allow clicking OnlyAllowCheckBoxClick with keyboard
Previously, attempts to click on a checkbox with "OnlyAllowCheckBoxClick" via the keyboard would've been consumed the same way that other regular clicks would be. Now I've added special handling to allow this to be done via the keyboard, since there would be no other way to do it via the keyboard.