Add more brush properties to TabControl
With these, the buttons in the TabControl should now be able to change their appearance based upon these burshes... but while it took for the close buttons in the TabDisplayItem, it didn't seem to work for the other buttons in the TabControl.

I'm going to have to figure out what's happening here.