Add DOVC, workaround bug with brushes in TabControl
Added in the DebugOutputValueConverter, which I was using to monitor changes to the brushes in the TabControl's template.

I made a workaround for the bugs I was having with the brushes not applying to the buttons in the TabControl. The workaround being explicitly finding the buttons and setting the properties. I don't like that it's come to this, but I can't figure out why the bindings aren't working the way they should.

Whatever. Anyway, I'll probably clean up a bit more and then work on releasing this thing.