Fix/rewrite brush handling for TabControl, TabDisplayItem
I rewrote how TabDisplayItem accesses and applies its brushes to the border component. As a part of that, a new IsHighlighted property was added. Now I use a WPF Style and it works instantly now.

I also fixed TabControl applying brushes to the menu of the TabDisplayItem. As it is right now, it applies the properties directly by using the already-stored MenuButton from the template. Eventually, I would like to not have to use all of these stored items and just use template bindings, but I guess that's a question for another release.

I'll be doing another once-over to make sure there's not any other lingering problems, and then I'll be ready to release!