Do not reset view settings on loading object
Now when PropertyList loads an object, the filter and view settings won't be reset (i.e. read-only properties, inherited properties). This was originally due to a limitation in the design of this control earlier on, but at this point, such limitations no longer exist.

Also, added in the FilterText property, to mirror what's offered in the 2.0 PropertyList, and inform people that property will be what's used going forward.