Add AllowAllNonModifiedKeysInShortcuts to KeyRegistry
This is a feature I wanted to add a while ago lol. With this, a KeyRegistry can now be used in situations where text input conflicts aren't a concern, and so keyboard shortcuts can be registered for all sorts of things!

This is kept as "false" by default to follow with existing behavior, and because I'd rather leave this false by default to prevent it accidentally causing a problem for someone down the road.

Also, I adjusted the registering code to exit early if a keyboard shortcut ends up not being registered due to the above things, to avoid raising an event for a shortcut that wasn't actually added.