Add StringSpinner control to Avalonia
Now this is the last spinner control that I needed to add! Yay!

Unlike the other spinners, most of this control's unique properties are direct properties. Not only did this make porting the control over from the WPF version easier, but it was kind of necessary since most of these properties are read-only.

In the future, I may revisit the code design of this control in both WPF and Avalonia, and we'll see what happens then!