Fix Items property for SelectPanel, seems to work
Fix the Items property in SelectPanel by using a weird workaround to get it to work the way Avalonia wants it to without compromising the actual design of the control.

There's still some aspects that I'm not sure if I've gotten quite right on this, but I think for the most part, this is here and working.