Create IPropertyEditorHost interface, implement it
The IPropertyEditorHost is used to replace the requirement for the ExperimentalPropertyList in the IPropertyEditor interface; this more generic interface can be used not just for the new PropertyList, but future controls I'll have in mind that'll use property editors.

At this point, I just put in a couple random-ish things to put in the interface, we'll see if I'll want to add in more as time goes.