Add XML comments to SsuiTheme, make it freezable
Add in XML comments to the various properties in the SssuTheme now, and also change it to inherit from Freezeable, rather than just DepednencyObject. This way, you can freeze a theme if you don't plan on changing it for your application. It also adds in built-in cloning support.

I still have more work to actually get this class all set up and ready to use; this is still just kind of the shell, and I need to fill in the insides.