Change CheckState to dependency property
Not sure why I didn't do this before when I was refactoring the CheckBox, but I'm doing it now!

I'll also be working to re-add the CheckBoxClick functionality, but I think that'll be happening primarily in the 2.0-ribbon branch for the moment; once that branch gets merged in (probably later this summer), then that'll be back in the main package too.