Add ThemedControl base class
This will be what I'll use as the base for a number of my SSUI classes in the future, rather than using just plain ol' Control. However, this isn't set up perfectly just yet, I'll be needing to make some more adjustments before this is ready to be implemented.

As a part of this, the SsuiTheme property that this introduces is actually built as an attached property with inheritance, which will result in a major change in how the rest of the library interacts with SsuiTheme, and also can reduce the amount of XAML needed to be written in the future too! It's unfortunately a bit of a tradeoff though, since it might introduce some minor performance and/or memory problems.

I'll give it a bit more of a think, but I do think this is the route I'll be going either way.