Implement SelectedTabChanged event for Ribbon
Didn't realize until now that this event actually didn't do anything lol. I changed the event to also be a DependencyPropertyChangedEventHandler event, rather than using the TabControl specific event handler we had before. This is better anyway, as you can also get an idea of what the old selected tab was as well.