Further work on Gallery control - mostly works!
The Gallery control is now mostly functional in the basic sense, but still has a couple critical bugs, and is also missing some extra features. I'll need to work further on fixing those, but for now, it's something!

As a part of this, I created three new commands to use with the Gallery for it to perform its functions.

I'll need to do further work on this in the future, including, most importantly, figuring out a way to update the visual tree of the Items when the menu is opened or closed. I think I'll probably see what Microsoft's Ribbon control does, since I don't know if I'd be able to recreate an equal or better solution.