Create ThemedContentControl
This is a ContentControl with the added features and properties of a ThemedControl.

I think I'm still having difficulties wrapping my head around this whole attached properties thing. I think in the long run, it's going to be a good idea to do this so that I can do inheritance, but there's a lot that I feel I'm still trying to learn or figure out, and there isn't as much good documentation or guidance about it online.