Initial work to improve Gallery experience
The issue right now is that the items in a Gallery can only have 1 visual parent, which, I mean, makes sense. So to have the items appear in two places at once, we'll have to secretly clear out the two containers' item sources and fill in one when needed, and fill in the other when that's needed.

So that's what I'm doing here. I haven't really tested this yet, we'll see if this works! I'll probably continue working on this in a future commit.