Added in WPF source code links to refer from later
Don't feel like actually putting in the work right now to try to work more on this attached property stuff, I'll save that for a little bit later. At least for now I put in links that I can pull up later to see how WPF implements the FlowDirection property.

FlowDirection is probably one of the few properties that is in the same situation that my SsuiTheme property is in, where it has one main control that hosts it, but then some other controls also implement it. So I'll see what I can learn from this!