Update FlatWindow properties, defaults, template
This encompasses a lot of various changes to FlatWindow, but primarily, a lot of work on renaming and updating the brushes.

A lot of the brushes were renamed and reorganized to better reflect what they are for and to avoid confusion. This will be a major change for anyone upgrading their projects where they customize the brushes of their windows. This 2.0 release will be the best time to make this change happen though, versus keeping the vague or sometimes incorrect names.

I removed the FlatWindow-specific Background, BorderBrush, and BorderThickness properties, so that we can instead use the underlying Control properties. I changed CaptionHeight to a dependency property, and removed the IsHighContrast property.

The default for CaptionButotnPadding is now 12,7,12,7.... this is closer to Windows 10 and 11's appearance versus previous versions, but still not quite there. I'm unsure if I want to move all the way to 15,7,15,7 now to match Windows 10/11, or keep this as it is until like 2.1 or 3.0 or something.

I also updated a lot of XML comments.

For the template, I moved the title bar to finally be its own grid. I also changed the caption buttons to no longer change their foreground when the window isn't active; instead, the buttons now go to 70% opacity. I also changed all the remaining FindAncestor bindings to instead use TemplatedParent.

This is a breaking change.