Add more visual properties to RectEdit
Added more properties to control the visuals and design of the RectEdit control. This includes increasing the space between the measure combo box and the rectangle area, and allowing the various text strings in the control to be changed.

I also removed the default height and width values for the control.