RectEdit: handle large values better; improve RectEditDialog
Set a maximum width for the spinners in RectEdit, to better prevent them from clipping outside of the control. Also changed the width and height of the RectEdit control in RectEdit to improve how it looks.

This fixes the issue of how it looks when setting the Rect to Rect.Empty.