Spinners: call UpdateUI after applying template
This way, we can make sure that the newly found text box actually has the correct value once it's set up. Template loading may not happen until after the Loaded event is called, so the existing process of calling ValidateValue once it's been loaded does not work, but it will at least make sure the spinner's value is validated at some part in the loading process, so it's still valuable to have.

This solves the issue that I was seeing in the RectEdit control.