Add ParticleEmitter control
This was something I developed in an internal repo for a personal project, since there wasn't an existing 2D particle emitter for WPF that I could find.

Well, I figure, if I couldn't find one, I may as well make mine open source! Overall, I'm not sure how often this control will actually be useful, but if someone ever wanted it, well, here it is.

This particle emitter is pretty decent but not perfect. It performs decently well and is thread-safe, but I haven't focused on performance or on adding on a wide selection of featuers. But for your basic usage, it should be good, and there are properties you can tweak to help it work better on low-power systems.