Remove IsDefault/IsDefaulted from FlatButton
Getting IsDefault to work property with my button would require a lot of reflection and such to get done. Which earlier me would not have had enough confidence to do, and present day me... well, I'm not opposed to doing, but I just don't want to lol.

I might come back to this later to properly implement, but I think the better option would be convincing the WPF team to accept a PR to make the necessary parts properly accessible via a protected method in ButtonBase, and I'm not sure how keen they'll be to make that happen.

For now, it's been made apparent since 1.9.7 of the broken state of these properties. While at the time I said I might explore what I can do about it in 2.0, the reality is that it's just not a priority for me.

As an alternative, I'd recommend either a developer use a standard WPF button for the default button and style it to look like a FlatButton, or continue with a FlatButton and add a KeyDown listener to the Window to listen for the Enter key.

This is a breaking change.