Create ThemedUserControl, improve property ownership
Created a ThemedUserControl class, which can be used where you have a user control but then want to add on the SSUI theming capabilities. Going to get a lot of use from this in my testing program, if nothing else.

Also changed the UserAccentTheme and UseLightBorder properties in ThemedContentControl, and the SsuiThemeChanged and SsuiThemeApplied events to instead just be added owners of the existing ThemedControl properties/events.

Also updated some XML comments.