MessageDialog: dependency properties, clean up code
Use dependency properties for a number of the properties of the window. Also did a bit of cleaning up code and improving XML comments. I did create a new overload for ShowDialog with less properties; I might add or remove some parameters on that one as needed. Just wanted a simpler overload to use.

For the XAML file, I cleaned up and rearranged a few things, and also put in some bindings for the dependency properties I had made.