Rearrange/improve LinkTextBlock code
I didn't realize that I haven't touched or improved this code at all. I know that I made some modifications to this in the Avalonia branch, but I'm not sure how far or what kind of changes I made.

Anyway, I rearranged the code and put it into regions. I also updated the XML comments for the dependency properties and Click routed event. I removed the override for the IsEnabled property, and instead just put in a listener for the underlying FrameworkElement's IsEnabled dependency property.

Finally, I changed the un-highlight code to not remove text decorations if UnderLineOnHighlight is set to false.

I'm kind of on the fence about making a 1.9.10, but if I do make it, this would be in it.