Add MessageDialog display in DialogsTest
I'm pretty sure I put this into the 1.9.7 branch version of this lol. When clicking on the text in the DialogsTest, it'll display a MessageDialog that displays the text.

This just makes a quick and easy way for me to test that MessageDialog works as expected.

Also, yes, this is what precipitated the LinkTextBlock changes in the previous commit lol.