Decouple TransparentBack from ApplyColorScheme
I think this will be more stable code, and this was also a prerequisite I had to do before I could apply SsuiTheme to the button. Which I'll have to do customly anyway, since FlatButton takes from ButtonBase, and not ContentControl.

Anyway, this might also fix issues I was having with the TabControl not able to apply its button theming correctly, possibly. I'll see if I get around to testing that later.