Improve Menu brush code, add MenuBorderBrush
Cleaned up and improved the code for declaring the brush properties, since apparently I just haven't done that.

As a part of this, I removed the overload for BorderBrush and instead renamed it to MenuBorderBrush, and made it specifically apply to the menus.

Instead, I now use BorderBrush and BorderThickness to describe the actual menu bar itself.

Also cleaned up a bit of other places with the template and the ColorScheme code.

This is a breaking change.