Add SsuiTheme, UseTitleBarThemeBrush to Menu
One more control now set with SsuiTheme! This wasn't that big an ordeal now since I now have what I did for FlatButton as the process I can follow.

As part of this, I created a new property UseTitleBarThemeBrush to control if this Menu should try to blend with the window's title bar or not.

Next, it'll be the ContextMenu time! And then I'll be ready to get more controls done!