Create MultiplierConverter classes for various types
Right now, I was only needing the ThicknessMultiplierConverter, but since I was going to go through the effort to make that one, I decided I may as well make it for a few other types too: double, CornerRadius, and Point.

These are actually pretty flexible, in which that they can take any number (int, float, etc.) or a parseable string as the value or parameter. It will attempt to convert that number or string to a double (or for the complex WPF types, will attempt to use its bespoke type converter or create a uniform object using that double), so many types of values can be put in and it will work. This is due to the cross-support of IConvertible.

Technically, as a result, this can be a way to provide a converter to convert doubles or strings into these WPF types, by setting the parameter to null, but not sure how useful that is lol.