Created PositonSelect to replace RelativePositionSelect
Not sure what exactly prompted me to decide to do this, but here we are!

This new version of this control has the same functionality and features, but is now built as a templated control, rather than being a UserControl. Some future adjustments will need to be made too to change SelectedPoint and SelectedX and SelectedY to be dependency properties.

I also need to add proper SsuiTheme support; it is a themed control, but right now it does not do anything with the theming.

In a future commit, I'll remove RelativePositionSelect once I've completed testing and made sure the new control is all good!

This is technically a breaking change for those converting from the old RelativePositionSelect to the new PositionSelect, since SelectedWidth and SelectedHeight were changed to SelectedX and SelectedY respectively.