Change SelectedPoint, X, and Y to dependency properties
This way, they can properly support binding and all the fun WPF features. To go along with this, I created helper methods and better organized the code I have.

Ideally now, everything should be all set and ready to go!

This also changes SelectedPositionChanged to a routed event. This is a breaking change.