Add SsuiTheme property/code to ContextMenu
Add in the SsuiTheme code and related things to the ContextMenu. Alongside this, I also fixed the dependency properties not having XML comments and cleaned and rearranged a bit of code.

I also removed the BorderBrush property I had defined; instead, we'll use the standard property that comes with Control. This is a breaking change.