Fix dependency property code in many places
I don't think I've gotten all of them, but I got to a lot of various places. I changed instances of "public static DependencyProperty" to add on "readonly", as is the general/best design for WPF controls.

I will need to update my PropertyBuilder program to actually have the "readonly" in there.

As a part of this, I also found some other spots to do things like fix XML comments and introduce the nameof constant.