Move value converters into the Converters folder
This just help keeps all these things organized not only for me but for users too. And this can also help with discoverability, as users of the library can go through the class list and see converters for items they didn't think they needed.

The most notable exception is keeping BindableChild and BindableChildConverter in the main namespace, as it felt a bit weird to lump those in with the proper value converters.