Change brushes in SelectableUserControl
I'm not the happiest with how the brushes are set up for SelectableUserControl, but I think I like this situation better than what I was doing before. The issue is that either this way or the previous way kind of breaks with standard WPF convention with how Background (and now Foreground) is handled, this way is at least just more visible about it.

With this, I've added on extra foreground brushes for the various states. I've also added on the SsuiTheme property and code, and so that should be set there too. I also now call UpdateBrushes in a few more situations.

Also, I added the SsuiTheme proeprty to the IClickSelectableControl interface finally.

I'm going to do a bit more thinking about how I want this situation with brushes and stuff handled for SelectableUserControl/SelectPanel and we'll go from there.

Getting SelectableUserControl updated, though, unlocks most everything else, I think, to go ahead and get updated with SsuiTheme now. So looking forward to converting more stuff soon!