Remove redundant properties in SelectPanel
Remove SelectPanel's own Background, BorderBrush, and BorderThickeness propreties, since those are defined in Control. Didn't realize I had these sitting there.

Also hooked up the ItemsTheme to be set when the SsuiTheme property is set. This is done via binding, and I hope the binding works, we'll see when I test all of this later on.