Change DoubleSpinner's Decimals back to an int
While ultimately the type chosen doesn't super matter, I think that "byte" may give the impression of it being a piece of data rather than an actual number that can be chosen (Based upon considering the other situations that bytes are often used in)

After some further reading, it seems the amount of memory and space that this property's value is going to take up will be the same size as an int. So I guess I may as well go with an int; and besides, other situations where there's a bounded integer property, others still go with int anyway.