Mostly reimplement OnlyAllowCheckBoxClick
Reimplemented a lot of the code that I had for CheckBox's OnlyAllowCheckBoxClick. It may not be 100% functioning correctly though, but for the most part, it seems to be there.

I'll need to spend some time later to make sure the visuals are actually working well, as I'm worried I broke something with that in this commit.