Remove unneeded GradientStop in GradientStopItem
Holding on to a GradientStop object isn't necessary, I wasn't even really doing anything with it! And holding onto it could actually cause a worse problem if the GradientStop object I'm storing was frozen at all.

Also added regions to organize the file.

This could be backported to a hypothetical 1.9.10.