Move MultilineStringInputDialog to Dialogs folder
Not quite sure why I put it in the Utils folder to begin with, although I suspect it's that I thought it wouldn't be that useful.

But meh, let's just put it here! As a result, more people will be able to see and hopefully find usefulness in this window!