Add ForegroundHighlightBrush to buttons (still buggy)
Add a new brush to the button controls, to change the color of the text/foreground while the control is highlighted... except I think it changes the border instead?? I'm going to have to investigate this further.

I also took this time to generally clean up a lot of aspects of the brush code and some template stuff with the FlatButton, MenuButton, and SplitButton. Less stuff being defined in the FlatButton's constructor now!

Also added a usage of SsuiThemeChanged in ButtonsTest, which I'll use to test this new brush.