Create SsuiTheme.CreateBinding method
Yay abstraction!

This will prove useful in the future for if I want to add support for setting a control's properties via a Style rather than binding the properties directly (as instead we can just bind the Style's values).

This could become part of a side quest to reduce the amount of code in ThemedControl (and other implementers) and instead move more of this logic back into SsuiTheme. Not a priority though, but we'll see.