Relayout, update MultilineStringInputDialog
Brought over a lot of the improvements from StringInputDialog and put them into MultilineStringInputDialog. This includes creating dependency properties for various things, and using grid rows to help organize the XAML.

Also changed the Value property in both dialogs to be a dependency property, I'm surprised it wasn't before.

Now both dialogs are much better suited for usage in WPF.