Update CharInputDialog layout, use SsuiTheme
Updated CharInputDialog to remove reliance on ColorScheme, Also subtlely changed the design of the window's rows and columns, and made sure that the tab order is correct and things won't get too wonky at weird sizes.

Also, changed a few of the common StringInputDialog properties to be dependency properites, but not much.