Set SsuiTheme to null by default, add nullability
Setting this property to be null by default will allow users to set their brushes and do things in XAML if they don't want to mess with the SsuiTheme.

As a part of this, I wanted to expose nullability in the .NET Core versions, which also meant adding on some null checks and some preprocessor if checks to make that happen.

During this, I also cleaned up a bit of code and also improved some XML comments to better describe the experience and use case with SsuiTheme.